Added class TournamentBuilder
